PURE-relation/requirements.txt
tqdm
allennlp==0.9.0
paddlepaddle-gpu==1.4.0
transformers==3.0.2
overrides==3.1.0
requests==2.25.1
